+++ src/components/Header/index.jsx
@@ -1,16 +1,23 @@
-import React from 'react';
-import PropTypes from 'prop-types';
+import React, { useState } from 'react';
 
+import { Header } from '@master-c8/commons';
 import Avatar from '@mui/material/Avatar';
-import { Box, MenuList, MenuItem, Link, ListItemIcon, ListItemText, Drawer, IconButton } from '@mui/material';
-import { Message, BriefCase, Notification, Blog, UserGroup, Website, Menu } from '@master-c8/icons';
+import { MenuList, MenuItem, Link, ListItemIcon, ListItemText, useMediaQuery, useTheme } from '@mui/material';
+import { Message, BriefCase, Notification, Blog, UserGroup, Website } from '@master-c8/icons';
 import { grey } from '@mui/material/colors';
 
-export const HeaderMobile = ({ openDrawer, setOpenDrawer }) => {
+import { HeaderMobile } from './HeaderMobile';
+
+export const HeaderEnterprise = () => {
+    const [openDrawer, setOpenDrawer] = useState(false);
+    const theme = useTheme();
+    const isMatch = useMediaQuery(theme.breakpoints.down('md'));
     return (
-        <Box>
-            <Drawer anchor="right" onClose={() => setOpenDrawer(false)} open={openDrawer}>
-                <MenuList sx={{ display: 'block' }}>
+        <Header className="header-menu" isLogged>
+            {isMatch ? (
+                <HeaderMobile openDrawer={openDrawer} setOpenDrawer={setOpenDrawer} />
+            ) : (
+                <MenuList sx={{ display: 'flex' }}>
                     <MenuItem>
                         <Link
                             sx={{
@@ -71,7 +78,7 @@ export const HeaderMobile = ({ openDrawer, setOpenDrawer }) => {
                                 alignItems: 'center',
                             }}
                             size="small"
-                            href="/interviews"
+                            href="/#/interviews"
                         >
                             <ListItemIcon>
                                 <UserGroup sx={{ color: grey[900] }} />
@@ -88,7 +95,7 @@ export const HeaderMobile = ({ openDrawer, setOpenDrawer }) => {
                                 alignItems: 'center',
                             }}
                             size="small"
-                            href="/interviews"
+                            href="/"
                         >
                             <ListItemIcon>
                                 <Blog sx={{ color: grey[900] }} />
@@ -105,7 +112,7 @@ export const HeaderMobile = ({ openDrawer, setOpenDrawer }) => {
                                 alignItems: 'center',
                             }}
                             size="small"
-                            href="/interviews"
+                            href="/"
                         >
                             <ListItemIcon>
                                 <Website sx={{ color: grey[900] }} />
@@ -113,21 +120,11 @@ export const HeaderMobile = ({ openDrawer, setOpenDrawer }) => {
                             <ListItemText sx={{ color: 'black' }}>Contacts</ListItemText>
                         </Link>
                     </MenuItem>
-                    <MenuItem sx={{ justifyContent: 'center' }}>
-                        <ListItemIcon sx={{ color: 'black' }} size="large">
-                            <Avatar sx={{ bgcolor: '#AE4EFF', width: 45, height: 45 }} alt="Profile" />
-                        </ListItemIcon>
-                    </MenuItem>
+                    <ListItemIcon sx={{ display: 'flex', alignItems: 'center', color: 'black' }} size="large">
+                        <Avatar sx={{ bgcolor: '#AE4EFF', width: 45, height: 45, ml: 3 }} alt="Profile" />
+                    </ListItemIcon>
                 </MenuList>
-            </Drawer>
-            <IconButton onClick={() => setOpenDrawer(!openDrawer)}>
-                <Menu />
-            </IconButton>
-        </Box>
+            )}
+        </Header>
     );
-};
-
-HeaderMobile.propTypes = {
-    openDrawer: PropTypes.bool.isRequired,
-    setOpenDrawer: PropTypes.func.isRequired,
 };
